Add new asl-backup-menu script
The new "asl-backup-menu" script helps to manage the backup/restore process on
AllStarLink nodes.  This script manages both local archives and remote archives
(on backup.allstarlink.org).

Note: many portions of the script were pulled (or derived) from the "asl-menu",
      "node-setup", "save-node", and "restore-node" scripts.

Next up will be to remove the backup/restore content from the other scripts and
exec this script to do the work.  It would also be good to consider removing the
save-node and restore-node scripts.