Add "Restart" option to the expert config menu (#47)
On the discussion lists I noticed people commenting about using the expert
configuration menus to update the configuration files.  While this is fine we
should remind folks that asterisk may need to be restarted.  Added a "Restart"
menu item to the menu and flag that this is needed whenever an [Asterisk]
configuration file is updated.

Also made some tweaks to have the menus remember the last selected item.