Menu improvements (#51)
- Added a more guided setup process for "new" nodes

For the initial (and any "new" nodes), we now provide a more guided setup process. We ask for the new node #, the node password, the type of node, the node callsign, etc. You no longer start off with a clean slate where you need to set every option.

- For duplex==0 "Half duplex with no telemetry (link)"

Nodes with no telemetry can actually be configured for "Full duplex" (controlled by "linktolink"). We are changing the menu to allow you to specify a "Half" or "Full" duplex configuration. Also, for half-duplex configurations we will also set "hangtime = 100".

- Some minor bug fixes and a handful of "correctness" changes
- Add more checks re: valid node #'s
- Do not enable statpost for private nodes
- Do not flag missing node [registration] passwords on private nodes
- If we have updated any of the configuration files then "sync" before exit
- Add a bit more polish :-)