node-setup : specify "node" when exec'ing "tune" menus (#62)
A change was made to the "simpleusb-tune-menu" and "radio-tune-menu"
commands that allows the "node" number to display/update to be specified
as a command line argument.

Here, we change the "node-setup" command to specify this new argument
when using the "Interface Tune CLI" menu option.  This will help to
ensure that any setting updates will be made to the already selected
node.