Add a new "Full-duplex node, that do not repeat audio" node type
When creating new nodes we prompt for the general type of node.  Our list
of commonly used node types did not include a full-duplex node that does
not repeat audio.  This change adds a new type that will start a node off
with defaults using the SimpleUSB channel driver and duplex=3.