node-setup: Allow first node to be "1999" (#82)
When setting up the first node you are prompted for the initial node number. We change the node included in the base configuration (1999) to the requested node. The gotcha is that we had been checking to ensure that the newly requested node number had not already been configured (you don't want two nodes with the same number). This check blocked anyone from trying to set the initial node number to "1999". This change allows any valid node number to be specified.