node-setup: also update node secret in IAX registrations
In general, we want everyone to use our HTTP based node registration.  But,
in some instances, HTTP registration is not available and one must switch
to using the older IAX registration.  Since the `node-setup` menu is already
setting/updating the node passwords for the former we can easily also
update the latter.