node-setup: allow comments on allowed/blocked nodes (#97)
We use an Asterisk database for the node allowlist/denylist.  The dialplan looks to the database keys to determine if a node should be allowed access or blocked from connecting.  The database values were being ignored but some have suggested that this could be a place to store a comment about why the node was added to a list.  This change adds support for using the database values as comments.