Ensure correct ownership/permissions of "/etc/asterisk/custom"
While investigating [Interface Tune CLI settings not being saved](https://community.allstarlink.org/t/23994),
we found that the "/etc/asterisk/custom" directory was not owned by
the "asterisk" user.  This blocked updates of the "tune" settings.

This change ensures that the ownership of the newly installed
"/etc/asterisk/custom" files and directories will be as expected.